src/application/useCases/sendNotification.spec.ts
import { InMemoryNotificationsRepository } from '../../../test/repositories/inMemoryNotifications';
import { SendNotification } from './sendNotification';

describe('Send notification', () => {
  it('should send a notification', async () => {
    const notificationsRepository = new InMemoryNotificationsRepository();
    const sendNotification = new SendNotification(notificationsRepository);

    const { notification } = await sendNotification.execute({
      category: 'social',
      content: 'This is a notification',
      recipientId: 'recipient-random-id',
    });

    expect(notification).toBeTruthy();
    expect(notificationsRepository.notifications).toHaveLength(1);
    expect(notificationsRepository.notifications[0]).toEqual(notification);
  });
});
